requirements.txt
python-gitlab==1.8.0
PyGithub==1.43.7
